cleanup storybook
