+++ src/plugin/_builtInStories/useGroupBy.stories.tsx
@@ -1,6 +1,6 @@
 import { withKnobs } from '@storybook/addon-knobs'
 import * as React from 'react'
-import { usePagination } from 'react-table'
+import { useGroupBy, useExpanded } from 'react-table'
 import styled from 'styled-components'
 
 import { FAKE_DATA, BASIC_COLUMNS } from '../../_fakeData/storyFakeData'
@@ -13,18 +13,32 @@ const Container = styled.div`
 `
 
 export default {
-  title: 'Plugins/usePagination [built in]',
+  title: 'Plugins/useGroupBy [built in]',
   decorators: [withKnobs],
 }
+
+const columns = [
+  {
+    Header: 'Group',
+    accessor: 'group',
+    Cell: ({ row: { groupByVal } }) => {
+      return groupByVal
+    },
+  },
+  ...BASIC_COLUMNS,
+]
 
 export const BasicExample = () => {
   const tableInstance = useTable<Faker.Card>(
     {
       data: FAKE_DATA,
-      columns: BASIC_COLUMNS,
-      pageSizeOptions: [5, 10, 15, 20],
+      columns,
+      initialState: {
+        groupBy: ['group'],
+      },
     },
-    usePagination
+    useGroupBy,
+    useExpanded
   )
 
   return (
