+++ src/plugin/useDensity/useDensity.stories.tsx
@@ -7,13 +7,15 @@ import { FAKE_DATA, BASIC_COLUMNS } from '../../_fakeData/storyFakeData'
 import { Table } from '../../Table'
 import { useTable } from '../../useTable'
 
+import { useDensity } from './useDensity'
+
 const Container = styled.div`
   height: 100vh;
   width: 100vw;
 `
 
 export default {
-  title: 'Plugins/usePagination [built in]',
+  title: 'Plugins/useDensity',
   decorators: [withKnobs],
 }
 
@@ -22,8 +24,24 @@ export const BasicExample = () => {
     {
       data: FAKE_DATA,
       columns: BASIC_COLUMNS,
-      pageSizeOptions: [5, 10, 15, 20],
     },
+    useDensity
+  )
+
+  return (
+    <Container>
+      <Table instance={tableInstance} />
+    </Container>
+  )
+}
+
+export const WithPagination = () => {
+  const tableInstance = useTable<Faker.Card>(
+    {
+      data: FAKE_DATA,
+      columns: BASIC_COLUMNS,
+    },
+    useDensity,
     usePagination
   )
