+++ src/plugin/useExpandAll/useExpandAll.stories.tsx
@@ -1,11 +1,13 @@
 import { withKnobs } from '@storybook/addon-knobs'
 import * as React from 'react'
-import { usePagination } from 'react-table'
+import { useGroupBy, useExpanded } from 'react-table'
 import styled from 'styled-components'
 
 import { FAKE_DATA, BASIC_COLUMNS } from '../../_fakeData/storyFakeData'
 import { Table } from '../../Table'
 import { useTable } from '../../useTable'
+
+import { useExpandAll } from './useExpandAll'
 
 const Container = styled.div`
   height: 100vh;
@@ -13,18 +15,33 @@ const Container = styled.div`
 `
 
 export default {
-  title: 'Plugins/usePagination [built in]',
+  title: 'Plugins/useExpandAll',
   decorators: [withKnobs],
 }
+
+const columns = [
+  {
+    Header: 'Group',
+    accessor: 'group',
+    Cell: ({ row: { groupByVal } }) => {
+      return groupByVal
+    },
+  },
+  ...BASIC_COLUMNS,
+]
 
 export const BasicExample = () => {
   const tableInstance = useTable<Faker.Card>(
     {
       data: FAKE_DATA,
-      columns: BASIC_COLUMNS,
-      pageSizeOptions: [5, 10, 15, 20],
+      columns,
+      initialState: {
+        groupBy: ['group'],
+      },
     },
-    usePagination
+    useGroupBy,
+    useExpanded,
+    useExpandAll
   )
 
   return (
